src/recommendation/requirements.txt
# Core dependencies
torch>=2.6.0
transformers>=4.36.0
pillow>=10.0.0
requests>=2.31.0

# LLaVA specific
# llava-torch>=1.6.0
accelerate==1.3.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Logging
python-json-logger>=2.0.7

# Image processing
numpy>=1.24.0
